+++ react-tenant-selector/src/components/LoginOptions/index.ts
@@ -2,3 +2,4 @@ export { default as LoginWithAzure } from './LoginWithAzure';
 export { default as LoginWithADFS } from './LoginWithADFS';
 export { default as LoginWithCognite } from './LoginWithCognite';
 export { default as LoginOrWrapper } from './LoginOrWrapper';
+export { default as LoginWithFakeIDP } from './LoginWithFakeIDP';