+++ react-tenant-selector/src/utils/index.ts
@@ -1,3 +1,3 @@
 export { default as config } from './config';
-export type { SidecarConfig } from './types';
+export type { SidecarConfig, FakeIdp } from './types';
 export * from './tenancy';